Calculate is now imported by default in Includes.php empty.php can now also unset the session and not the mysql processReceipt now works, only print and document generation is not done yet. We now use a new system to generate a receiptId using UNIX_TIMESTAMP() from MySQL - 10 years in seconds + randon between 0 and 300 Price change modal now looks much more clearly presented and should be implemented into the item editor as well (#14) Total price for receipt is now calculated using the new Calculate::getReceiptTotal function

+++ src/receipt/empty.php
@@ -10,19 +10,22 @@ if (isset($_GET['receiptId']))
     $_SESSION['receipt']['items'] = null;
     $_SESSION['receipt']['status'] = 'closed';
     $_SESSION['receipt']['customer'] = null;
-    
-    $db = new mysqli($config['SQL_HOST'], $config['SQL_USER'], $config['SQL_PASS'], $config['SQL_DB']);
 
-    if($db->connect_errno > 0)
+    if (!isset($_GET['save']))
     {
-        die('Unable to connect to database [' . $db->connect_error . ']');
-    }
+        $db = new mysqli($config['SQL_HOST'], $config['SQL_USER'], $config['SQL_PASS'], $config['SQL_DB']);
 
-    $sql = "DELETE FROM receipt WHERE receiptId='" . $_GET['receiptId'] . "'";
+        if($db->connect_errno > 0)
+        {
+            die('Unable to connect to database [' . $db->connect_error . ']');
+        }
 
-    if(!$result = $db->query($sql))
-    {
-        die('There was an error running the query [' . $db->error . ']');
+        $sql = "DELETE FROM receipt WHERE receiptId='" . $_GET['receiptId'] . "'";
+
+        if(!$result = $db->query($sql))
+        {
+            die('There was an error running the query [' . $db->error . ']');
+        }
     }
 }
 ?>
